pyproject.toml
[project]
name = "verbenform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2026.3",
    "Django==6.0.1",
    "django-crispy-forms>=2.6",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
